requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.3
django-cors-headers==2.1.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.10
openapi-codec==1.3.2
pytz==2017.2
requests==2.18.3
simplejson==3.11.1
uritemplate==3.0.0
urllib3==1.22
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
cffi==1.10.0
constantly==15.1.0
cryptography==2.0.2
cssselect==1.0.1
hyperlink==17.3.0
idna==2.5
incremental==17.5.0
lxml==3.8.0
parsel==1.2.0
pyasn1==0.3.1
pyasn1-modules==0.0.10
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.2.0
queuelib==1.4.2
Scrapy==1.4.0
service-identity==17.0.0
six==1.10.0
Twisted==17.5.0
w3lib==1.17.0
zope.interface==4.4.2